requirements.txt
python-telegram-bot==21.6
pytesseract==0.3.10
Pillow==10.4.0
python-dotenv==1.0.1